+++ core/include/keyman/meson.build
@@ -24,8 +24,10 @@ configure_file(
 )
 
 install_headers('keyman_core_api.h',
+                'keyman_core_api_actions.h',
                 'keyman_core_api_bits.h',
                 'keyman_core_api_consts.h',
+                'keyman_core_api_context.h',
                 'keyman_core_api_debug.h',
                 'keyman_core_api_vkeys.h',
                 join_paths(meson.current_build_dir(), 'keyman_core_api_version.h'),
