+++ core/src/keyman_core.h
@@ -1,5 +1,6 @@
 #pragma once
 
 #include <keyman/keyman_core_api.h>
+#include <keyman/keyman_core_api_actions.h>
 #include <keyman/keyman_core_api_context.h>
 #include <keyman/keyman_core_api_debug.h>
